refactor(RHTAPWATCH-116): `querygen` Go program
Create the `querygen` Go program for generating query strings for Splunk
and use that in `fetch-uj-records.sh` instead of the bash code we had
for this before.

Doing this as a first step to switching to having multiple queryes and
more complex ones. Using Go for this allows for more complex query
generating logic then wnat we can easily do in `bash`.