chore(RHTAPWATCH-361): go test framework
Adds the containerfixture script that will be served as a go
test framework, with the ability to build and deploy containerized
services for us to perform tests on with real data.

The Splunk container specifically is injected with indexes and logs
for the tests, therefor requiring specific verification before
running the tests.

Currently the script intentionaly does not use the podman/bindings
library in attempt to avoid adding too many uneeded dependencies.

Signed-off-by: Omer <oamsalem@redhat.com>