feat(RHTAPWATCH-301): Map files
Instead of running the map generation scripts, `splunk-to-segment.sh`
now reads the mapping data from files.

The idea here being that when running in a cluster, those mapping files
would be saved as config maps and mounted on the POD running the scripts

Here are commands for creating said config maps:
```
oc create configmap ws-map \
  --from-file=ws-map.json=<(KUBECONFIG=some_cfg get-workspace-map.sh)
oc create configmap uid-map \
  --from-file=uid-map.json=<(KUBECONFIG=some_cfg get-uid-map.sh)
```
Where "oc" should be configured to connect to the cluster where the maps
are to be created, and `some_cfg` is a kube config file foe connecting
to the RHTAP clusters.

Signed-off-by: Barak Korren <bkorren@redhat.com>