fix(RHTAPWATCH-522): container builds flakiness
The Splunk container had a few issues that caused flakiness:

- The second stage of the dockerfile did not always finished copying
the DB (digested logs) before finishing the build process, leading
to failures.
- `podman kube play` also seems to cause instability when building
the image.

This change solve those issues by:
1. Add a 10 second wait time between copying the Splunk DB
and finishing the build process; for now, it seems to be enough
time for the process to complete consecutively.
2. Change the test framework so instead of relaying on `podman kube play`
for both building and running the container, the framework will build
using `podman build`, and run the container with `podman kube play`
using the built image.

Signed-off-by: Omer <oamsalem@redhat.com>