chore: error-handling for log indexing
We encountered an edge-case in which the curl command
that initiate the log indexing for Splunk silently fail.
As the looping check we have later on only checks for
indexing that already started, therefore ignoring what
failed to begin.

This change will attempt to retry the indexing initiation
for three times, until failing the script, given an
error will occur.

Signed-off-by: Omer <oamsalem@redhat.com>