chore: remove description title from PR
Removes the "Description" title from the pull request
template.
The idea beyond this change is that we are to expect
description to be a part of the commit message anyway,
meaning it'll be automatically added to the beginning
of the pull request description anyway.

having the description title there just adds the need
to manually move the text, which don't seem to happen
much, creating confusing PR description.

Signed-off-by: Omer <oamsalem@redhat.com>