chore(RHTAPWATCH-208): Run GO unit tests for each opened PR
Adds a new GH actions workflow that runs the GO unit tests
on each opened pull request.

This excludes the kwok and splunk packages tests, as those
include deployment of containers, which we don't want the
GH actions container to deal with.

This is also made into a new PR-only workflow to allow for
the future addition of codecov.

Signed-off-by: Omer <oamsalem@redhat.com>